# docs/swagger.yaml
definitions:
...
  model.UserJson:
    properties:
      captcha:
        type: string
      email:
        type: string
      name:
        type: string
      passWord:
        type: string
      role:
        type: integer
      userName:
        type: string
    type: object
  model.WellJson:
    properties:
      ID:
        type: integer
      address:
        type: string
      affiliation:
        type: string
      averageDailyProduction:
        type: integer
      boreholeSize:
        type: integer
      construction:
        type: integer
      depth:
        type: integer
      finishTime:
        type: integer
      life:
        type: integer
      note:
        type: string
      wellName:
        type: string
      wellType:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/user/check:
    get:
      parameters:
      - description: 用户名
        in: query
        name: userName
        required: true
        type: string
      responses: {}
      summary: 检验用户名是否重复
      tags:
      - User
  /api/v1/user/d:
    post:
      parameters:
      - description: uid
        in: formData
        name: uid
        required: true
        type: string
      responses: {}
      summary: 移除用户
      tags:
      - User
  /api/v1/user/info:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 获取当前登录用户信息
      tags:
      - User
  /api/v1/user/login:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: userName
        required: true
        type: string
      - description: 密码
        in: formData
        name: passWord
        required: true
        type: string
      responses: {}
      summary: 登录
      tags:
      - User
  /api/v1/user/new:
    post:
      parameters:
      - description: 用户信息
        in: body
        name: ud
        required: true
        schema:
          $ref: '#/definitions/model.UserJson'
      responses: {}
      summary: 用户注册
      tags:
      - User
  /api/v1/user/rp:
    post:
      parameters:
      - description: uid
        in: formData
        name: uid
        required: true
        type: string
      responses: {}
      summary: 重置用户密码
      tags:
      - User
  /api/v1/user/ru:
    post:
      parameters:
      - description: 用户名称
        in: formData
        name: name
        type: string
      - description: 旧密码
        in: formData
        name: oldPassWord
        type: string
      - description: 新密码
        in: formData
        name: newPassWord
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      summary: 用户修改信息
      tags:
      - User
  /api/v1/user/v:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: userName
        required: true
        type: string
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      responses: {}
      summary: 发送验证码
      tags:
      - User
  /api/v1/well/{id}:
    get:
      parameters:
      - description: 井ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 获取井的详细信息
      tags:
      - Well
...
  /api/v1/well/d:
    post:
      parameters:
      - description: 井ID
        in: formData
        name: wellID
        required: true
        type: integer
      responses: {}
      summary: 删除井信息
      tags:
      - Well
  /api/v1/well/new:
    post:
      parameters:
      - description: 井信息
        in: body
        name: wd
        required: true
        schema:
          $ref: '#/definitions/model.WellJson'
      responses: {}
      summary: 新增井信息
      tags:
      - Well
  /api/v1/well/rw:
    post:
      parameters:
      - description: 井信息
        in: body
        name: wd
        required: true
        schema:
          $ref: '#/definitions/model.WellJson'
      responses: {}
      summary: 编辑井信息
      tags:
      - Well
swagger: "2.0"